requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyTelegramBotAPI==4.9.0
requests==2.28.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
TelegramBotAPI==0.3.2
urllib3==1.26.14
